# deprecated/go/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "超低资源消耗超快的离线翻译服务器 API",
        "title": "MTranServer API",
        "termsOfService": "https://github.com/xxnuo/MTranServer",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/xxnuo/MTranServer/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/xxnuo/MTranServer/blob/main/LICENSE"
        },
        "version": "3.0.0"
    },
    "host": "localhost:8989",
    "basePath": "/",
    "paths": {
...
        "/google/language/translate/v2": {
            "post": {
                "description": "兼容 Google Translate API v2 的翻译接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "插件"
                ],
                "summary": "Google 翻译兼容接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "key",
                        "in": "query"
                    },
                    {
                        "description": "Google 翻译请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleTranslateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleTranslateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/google/translate_a/single": {
            "get": {
                "description": "兼容 Google translate_a/single 的翻译接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "插件"
                ],
                "summary": "Google translate_a/single 兼容接口",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gtx",
                        "description": "客户端类型",
                        "name": "client",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "en",
                        "description": "源语言代码",
                        "name": "sl",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "zh-CN",
                        "description": "目标语言代码",
                        "name": "tl",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "t",
                        "description": "数据类型",
                        "name": "dt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Hello, world!",
                        "description": "待翻译文本",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/translate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyQuery": []
                    }
                ],
                "description": "翻译单个文本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译"
                ],
                "summary": "单文本翻译",
                "parameters": [
                    {
                        "description": "翻译请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TranslateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TranslateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translate/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyQuery": []
                    }
                ],
                "description": "批量翻译多个文本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译"
                ],
                "summary": "批量翻译",
                "parameters": [
                    {
                        "description": "批量翻译请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TranslateBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TranslateBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.GoogleTranslateRequest": {
            "type": "object",
            "required": [
                "q",
                "source",
                "target"
            ],
            "properties": {
                "format": {
                    "type": "string",
                    "example": "text"
                },
                "q": {
                    "type": "string",
                    "example": "The Great Pyramid of Giza"
                },
                "source": {
                    "type": "string",
                    "example": "en"
                },
                "target": {
                    "type": "string",
                    "example": "zh-Hans"
                }
            }
        },
        "handlers.GoogleTranslateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "translations": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "translatedText": {
                                        "type": "string",
                                        "example": "吉萨大金字塔"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "handlers.TranslateBatchRequest": {
            "type": "object",
            "required": [
                "from",
                "texts",
                "to"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "en"
                },
                "html": {
                    "type": "boolean",
                    "example": false
                },
                "texts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Hello",
                        " world!",
                        "Good morning!"
                    ]
                },
                "to": {
                    "type": "string",
                    "example": "zh-Hans"
                }
            }
        },
        "handlers.TranslateBatchResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "你好，世界！",
                        "早上好！"
                    ]
                }
            }
        },
        "handlers.TranslateRequest": {
            "type": "object",
            "required": [
                "from",
                "text",
                "to"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "en"
                },
                "html": {
                    "type": "boolean",
                    "example": false
                },
                "text": {
                    "type": "string",
                    "example": "Hello, world!"
                },
                "to": {
                    "type": "string",
                    "example": "zh-Hans"
                }
            }
        },
        "handlers.TranslateResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string",
                    "example": "你好，世界！"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyQuery": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}